Rename Email Alert to Notification
