+++ frappe/email/doctype/notification/notification.json
@@ -1062,7 +1062,7 @@
  "istable": 0, 
  "max_attachments": 0, 
  "menu_index": 0, 
- "modified": "2018-05-31 15:43:58.646159", 
+ "modified": "2018-05-31 15:44:23.082691", 
  "modified_by": "Administrator", 
  "module": "Email", 
  "name": "Notification",
