+++ frappe/hooks.py
@@ -158,7 +158,7 @@ scheduler_events = {
 		"frappe.core.doctype.error_log.error_log.set_old_logs_as_seen",
 		"frappe.desk.doctype.event.event.send_event_digest",
 		"frappe.sessions.clear_expired_sessions",
-		"frappe.email.doctype.email_alert.email_alert.trigger_daily_alerts",
+		"frappe.email.doctype.notification.notification.trigger_daily_alerts",
 		"frappe.async.remove_old_task_logs",
 		"frappe.utils.scheduler.disable_scheduler_on_expiry",
 		"frappe.utils.scheduler.restrict_scheduler_events_if_dormant",
