+++ frappe/public/js/frappe/misc/help_links.js
@@ -33,8 +33,8 @@ frappe.help.help_links['List/Email Account'] = [
 	{ label: 'Email Account', url: 'https://frappe.github.io/erpnext/user/manual/en/setting-up/email/email-account' },
 ]
 
-frappe.help.help_links['List/Email Alert'] = [
-	{ label: 'Email Alert', url: 'https://frappe.github.io/erpnext/user/manual/en/setting-up/email/email-alerts' },
+frappe.help.help_links['List/Notification'] = [
+	{ label: 'Notification', url: 'https://frappe.github.io/erpnext/user/manual/en/setting-up/email/email-alerts' },
 ]
 
 frappe.help.help_links['Form/Print Settings'] = [
